Move auth functions to the auth helpers
